Enforce stricter compilation in lib CMakeLists

+++ lib/include/private/mimir/framelimit.hpp
@@ -1,9 +1,11 @@
 #pragma once
 
+#include <cstdint> // int64_t
+
 namespace mimir
 {
 
-void frameStall(unsigned target_rate);
-int getTargetFrameTime(bool enable, int target_fps);
+void frameStall(int64_t target_rate);
+int64_t getTargetFrameTime(bool enable, int target_fps);
 
 } // namespace mimir
